fixed script name collision with classes
